Move INesHeader to new file
